Change the puppet catalog diff tool to be a Puppet Face
This will ease the integration. To run the tool on two catalogs:

  puppet catalog diff catalog1.yaml catalog2.pson

This change introduces no code change, only bits moved around.
